+++ CHANGELOG.md
@@ -128,8 +128,9 @@ When watching fields `[up-watch]`, `[up-autosubmit]`, `[up-switch]` or `[up-vali
 - Navigation containers can now match the current location of other layers by setting an `[up-layer]` attribute
 - Listeners to `up:request:load` can now [inspect or mutate request options](/up:request:load#changing-requests)
   before it is sent. This has been possible before, but was never documented.
--  ⚠️ Renamed property up.RenderResult#options to #renderOptions
--  ⚠️ Renamed property up.RenderJob#options to #renderOptions
+-  ⚠️ Renamed property `up.RenderResult#options` to `#renderOptions`
+-  ⚠️ Renamed property `up.RenderJob#options` to `#renderOptions`
+- Added a property `up.RendeResult#ok`. It indicated whether the render pass has rendered a [successful response](/failed-responses).
 - iOS: Long-pressing a link to open the context menu will no longer follow an `[up-instant]` link (#271)
 - iOS: Long-pressing a link will no longer emit an `up:click` event (#721)
 - Fix a bug where history wasn't updated when a response contains comments before the `<!DOCTYPE>` or `<html>` tag (fixes #726)
@@ -143,6 +144,7 @@ When watching fields `[up-watch]`, `[up-autosubmit]`, `[up-switch]` or `[up-vali
 - It is now guaranteed that `up:fragment:inserted` is emitted after compilation.
 - `up.element.subtree()` now prevents redundant array allocations.
 - Requests now clear out their `{ bindLayer }` property after loading, allowing layer objects to be garbage-collected.
+- Fix a bug where `up.RenderResult#renderOptions` was sometimes `undefined`.
 
 
 ### Developer experience